+++ apps/calmar/CHANGELOG.md
@@ -1,5 +1,12 @@
 # calmar
 
+## 0.4.20260924
+
+### Patch Changes
+
+- Updated dependencies [39a9ea5]
+  - @neurodesk/webapp-components@0.4.5
+
 ## 0.4.20260923
 
 ### Patch Changes